pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scanllm"
version = "2.0.0"
description = "AI Dependency Intelligence — scan, govern, and visualize AI in your code."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "ScanLLM", email = "hello@scanllm.ai"},
]
keywords = ["ai", "llm", "security", "scanning", "sbom", "ai-bom", "owasp", "governance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "networkx>=3.0",
    "gitpython>=3.1",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.25.0",
]
reports = [
    "xhtml2pdf>=0.2.11",
    "jinja2>=3.1",
]
watch = [
    "watchdog>=4.0",
]
all = [
    "scanllm[server,reports,watch]",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[project.scripts]
scanllm = "cli.main:app"

[project.urls]
Homepage = "https://scanllm.ai"
Repository = "https://github.com/isunilsharma/scanllm"
Documentation = "https://docs.scanllm.ai"
Issues = "https://github.com/isunilsharma/scanllm/issues"

[tool.setuptools.packages.find]
include = ["core*", "cli*"]

[tool.setuptools.package-data]
"core.signatures" = ["*.yaml"]
"core.scoring" = ["*.yaml"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
